refactor(checker-tests): split multi-file helpers out of test_utils
test_utils.rs reached 2017 lines (over the §19 2000-LOC cap) after adding
check_multi_file_with_global_index. Extract the four check_multi_file*
pipeline helpers into a test_utils/multi_file.rs submodule (re-exported, so
call sites are unchanged). test_utils.rs: 2017 -> 1749 lines.

AgentName: M1Opus48
